requirements.txt
altair==5.5.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
comm==0.2.3
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
dateparser==1.2.2
decorator==5.2.1
dill==0.4.0
executing==2.2.1
finnhub-python==2.4.25
fonttools==4.60.1
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
idna==3.10
imageio==2.37.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
llvmlite==0.44.0
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.6.0
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pandas-ta==0.4.71b0
parso==0.8.5
peewee==3.18.2
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.1
prompt_toolkit==3.0.52
protobuf==6.32.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pybit==5.12.0
pycoingecko==3.2.0
pycparser==2.23
pycryptodome==3.23.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rpds-py==0.27.1
schedule==1.2.2
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
smmap==5.0.2
soupsieve==2.8
SQLAlchemy==2.0.43
stack-data==0.6.3
streamlit==1.50.0
streamlit-autorefresh==1.0.1
structlog==25.4.0
ta==0.11.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vectorbt==0.28.1
wcwidth==0.2.14
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
yfinance==0.2.66
